+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Breaking changes
 New features
 ------------
 
+* String interpolation support is added via the `` `... ${`` ... ``} ...` `` syntax.
 * `FileModuleResolver` resolves relative paths under the parent path (i.e. the path holding the script that does the loading). This allows seamless cross-loading of scripts from a directory hierarchy instead of having all relative paths load from the current working directory.